+++ Project Code/java-pommerman-master/src/groupO/RLLearner.java
@@ -1,12 +1,7 @@
-package players.RLplayer;
+package groupO;
 
 import core.GameState;
-import players.Player;
-import players.heuristics.AdvancedHeuristic;
-import players.heuristics.CustomHeuristic;
-import players.mcts.SingleTreeNode;
 import utils.ElapsedCpuTimer;
-import utils.Pair;
 import utils.Types;
 import utils.Vector2d;
 
@@ -141,64 +136,12 @@ public class   RLLearner {
         }
 
 
-        //for time limit
-            //Evaluate current state
 
-            // (loop) evaluate coridinates in the line of sight
-            //Move to best value based on current and cordinates around it
-            //if every value is same choose random state
-                //calculate q val based on action and store it
 
 
 
 
     }
 
-//    //calculate reward for state here
-//    private double calculateQ(GameState state) {
-//        int index = 0;
-//        int numIters = 0;
-//        boolean stop = false;
-//        int acumTimeTaken = 0;
-//        int avgTimeTaken = 0;
-//        long remaining = 0;
-//        int remainingLimit = 5;
-//        int fmCallsCount = 0;
-//        double newQ = Double.MIN_VALUE;
-//        int actionToPick = 0;
-//
-//
-//        ElapsedCpuTimer ect = new ElapsedCpuTimer();
-//        ect.setMaxTimeMillis(40);
-//
-//
-//        while (!stop) {
-//           // states.add(state);
-//            newQ = policy.getPolicyFromState(state);
-//            ElapsedCpuTimer elapsedTimerIteration = new ElapsedCpuTimer();
-//         //   qValues.add(newQ);
-//
-//            //TODO
-//
-//            if (params.stop_type == params.STOP_TIME) {
-//                numIters++;
-//                acumTimeTaken += (elapsedTimerIteration.elapsedMillis());
-//                avgTimeTaken = acumTimeTaken / numIters;
-//                remaining = ect.remainingTimeMillis();
-//                stop = remaining <= 2 * avgTimeTaken || remaining <= remainingLimit;
-//            } else if (params.stop_type == params.STOP_ITERATIONS) {
-//                numIters++;
-//                stop = numIters >= params.num_iterations;
-//            } else if (params.stop_type == params.STOP_FMCALLS) {
-//                fmCallsCount += params.rollout_depth;
-//                stop = (fmCallsCount + params.rollout_depth) > params.num_fmcalls;
-//            }
-//
-//
-//        }
-//        return newQ;
-//
-//
-//
-//    }
+
 }
