+++ Project Code/java-pommerman-master/src/groupO/RLParams.java
@@ -1,4 +1,4 @@
-package players.RLplayer;
+package groupO;
 
 import players.optimisers.ParameterSet;
 import utils.Pair;
